--- src/components/Modal.tsx
@@ -1,18 +1,0 @@
-import Box from '@mui/material/Box';
-import Modal from '@mui/material/Modal';
-import { TypePropsModal } from '../utils/types/types';
-
-const BasicModal: React.FC<TypePropsModal> = ({ open, handleClose, content }) => (
-  <div>
-    <Modal
-      open={open}
-      onClose={handleClose}
-      aria-labelledby="modal-modal-title"
-      aria-describedby="modal-modal-description"
-    >
-      <Box className="modal">{content}</Box>
-    </Modal>
-  </div>
-);
-
-export default BasicModal;
