+++ src/core/Layout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom';
-import Footer from '../components/Footer';
-import Header from '../components/Header';
+import Footer from '../components/Footer/Footer';
+import Header from '../components/Header/Header';
 
 export default function Layout() {
   return (
