+++ src/style/style.scss
@@ -1,3 +1,2 @@
 @import './base/index.css';
 @import './base/App.scss';
-@import './components/Modal.scss';
